requirements.txt
Flask 
requests 
pytz 
Pillow 
waitress
yfinance 
numpy 
opencv-python-headless
adafruit-blinka 
adafruit-circuitpython-bme680 
spidev 
smbus 
gpiozero
cairosvg
lgpio
rpi.gpio
pigpio
colorama
smbus2
adafruit-circuitpython-bme280
pimoroni-bme280
bme680
